+++ src/main/java/com/CrudSpringRest/resources/ClienteResource.java
@@ -1,4 +1,5 @@
 package com.CrudSpringRest.resources;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,7 +18,7 @@ public class ClienteResource {
 	private ClienteService service;
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.GET)
-	public ResponseEntity<?> find(@PathVariable Long id) {
+	public ResponseEntity<?> find(@PathVariable Integer id) {
 		Cliente obj = service.buscar(id);
 		return ResponseEntity.ok().body(obj);
 	}
